👳🏽 Changed Double to Float32. Added documentation for Application and ApplicationDriver. Created skeleton code for Matrix

+++ apptest/apptest/ios/iosGraphicsDriver.swift
@@ -25,7 +25,7 @@ class iosGraphicsDriver: GraphicsDeviceDriver {
         return      mtlDevice
     }
 
-    func            createGraphicsSurface   (with initialSize: Vector2, _ dpi: Double, to target: Any) -> Any {
+    func            createGraphicsSurface   (with initialSize: Vector2, _ dpi: Float32, to target: Any) -> Any {
 
         guard let   mtlDevice       = self.mtlDevice        else { FastFail.instance.crash(with: .unexpectedPayload) }
 
@@ -34,20 +34,20 @@ class iosGraphicsDriver: GraphicsDeviceDriver {
         guard let   cametalLayer    = vc.rootViewController?.view.layer as? CAMetalLayer else { FastFail.instance.crash(with: .unexpectedPayload) }
 
         cametalLayer.device         = mtlDevice
-        cametalLayer.drawableSize   = .init(width: initialSize.x, height: initialSize.y)
+        cametalLayer.drawableSize   = .init(width: CGFloat(initialSize.x), height: CGFloat(initialSize.y))
         cametalLayer.pixelFormat    = .bgra8Unorm
         cametalLayer.isOpaque       = true
 
         return  cametalLayer
     }
 
-    func            resizeGraphicsSurface   (_ graphicsSurface: GraphicsSurface, to newSize: Vector2, _ newDpi: Double) {
+    func            resizeGraphicsSurface   (_ graphicsSurface: GraphicsSurface, to newSize: Vector2, _ newDpi: Float32) {
 
         guard let caMetalLayer  = graphicsSurface.backingSurface as? CAMetalLayer else { FastFail.instance.crash(with: .unexpectedPayload) }
 
-        let scale               = Double(UIScreen.main.scale)
+        let scale               = Float32(UIScreen.main.scale)
 
-        caMetalLayer    .drawableSize   = .init(width: newSize.x * scale, height: newSize.y * scale)
+        caMetalLayer    .drawableSize   = .init(width: CGFloat(newSize.x * scale), height: CGFloat(newSize.y * scale))
         graphicsSurface .backingSize    = newSize * scale
     }
 }
